requirements.txt
# Framework principal
Flask==3.0.3
Werkzeug==3.0.3

# Base de datos
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Autenticacion
Flask-JWT-Extended==4.6.0

# CORS
Flask-Cors==4.0.1

# Email
Flask-Mail==0.10.0

# Variables de entorno
python-dotenv==1.0.1

# Servidor de produccion
gunicorn==22.0.0
